feat: added user query and refactored task
