+++ schema.js
@@ -1,29 +1,28 @@
-import {readFileSync} from 'fs';
+import { readFileSync } from 'fs';
+import { findUserById, getUsers } from './user.js';
+import { getTaskById, getTasks } from './task/index.js';
 
 const typeDefs = readFileSync('./schema.graphql', { encoding: 'utf-8' });
 
 const resolvers = {
   Query: {
     tasks() {
-      return tasks
+      return getTasks();
     },
     task(_root, args) {
-      return tasks.find(task => task.id == args.id)
+      return getTaskById(args.id)
+    },
+    users(){
+      return getUsers();
+    },
+    async loggedUser(_root, _args, contextValue){
+      if (!contextValue.userId){
+        return null;
+      }
+      const { username, id } = await findUserById(contextValue.userId);
+      return { username, id };
     }
   }
 }
-
-const tasks = [
-  {
-    id: "1",
-    title: "Hello world",
-    description: "Task description"
-  },
-  {
-    id: "2",
-    title: "Hello world 2",
-    description: "Task description 2"
-  }
-]
 
 export { typeDefs, resolvers };
